test(providers): move the catalogue-cache tests beside what they test
They landed in the provider-resolution invariants file, whose scope is name
resolution and source-shape tripwires. Caching is neither, and the catalogue file
already imports the module under test. Raised in review.
